elish : Add basic software that can be used in va version, such as music player, chrome, JamesDsp, Google Camera
Since the bug that caused the desktop to crash due to screen rotation was fixed in QPR2, and the small bug that the search in the settings was prone to crashing, we have to say goodbye to RisingOS-Revived. Thank you for your company and support.

RisingOS-Revived QPR1 has been suspended. This does not mean that QPR2 will not be built. Please rest assured that I am also looking for suitable new projects for everyone to play and communicate.

You can find me via
QQ：1634033479
QQ Group :820085913